feat(action): add install-and-archive GitHub Action (#138)
This commit introduces a new GitHub Action for installing packages and
archiving installed files. The action allows users to specify
installation scripts, archive paths, and directories to monitor for
changes. Detailed usage instructions are provided in the README.

--- commit logs ---
* feat(action): add install-and-archive GitHub Action
This commit introduces a new GitHub Action for installing packages and
archiving installed files. The action allows users to specify
installation scripts, archive paths, and directories to monitor for
changes. Detailed usage instructions are provided in the README.