fix(docs): update action references in README files (#142)
Updated the action references in the README files for
`actions-install-and-archive` and `actions-use-apt-tools` to point to
the new repository location `shiron-dev/actions` instead of
`tecolicom/actions`. This change ensures that users are directed to the
correct source for the actions.

--- commit logs ---
* fix(docs): update action references in README files
Updated the action references in the README files for
`actions-install-and-archive` and `actions-use-apt-tools` to point to
the new repository location `shiron-dev/actions` instead of
`tecolicom/actions`. This change ensures that users are directed to the
correct source for the actions.